start/requirements.txt
Flask>=0.12
#google-cloud>=0.27.0
#google-cloud-core>=0.27.1
#google-cloud-datastore>=1.3.0
google-cloud-ndb==1.7.1
jsonschema==2.6.0
pyyaml==5.3.1